Made clear distinction between server and port forwarder
